fix(cuda): keep Field residency compile-check out of Catch2
The CUDA residency TU used a namespace-scope constructor, so loading
openpfc-tests ran CUDA during CTest discovery on GPU-less runners.
It is now an uncalled function in an OBJECT library, built when CUDA
is on, never linked into openpfc-tests.
