+++ tests/unit/kernel/data/CMakeLists.txt
@@ -4,3 +4,5 @@
 # Kernel data layer unit tests
 file(GLOB DATA_UNIT_TEST_SOURCES CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/test_*.cpp)
 target_sources(openpfc-tests PRIVATE ${DATA_UNIT_TEST_SOURCES})
+# cuda_field_residency_compile.cpp is an OBJECT library in
+# cmake/LibraryConfiguration.cmake; it is not part of openpfc-tests.
